revc2/requirements.txt
websockets
pillow